requirements.txt
requests
elasticsearch
pandas
numpy
matplotlib
urllib3
flask
tqdm
python-dotenv